# Use Castle to generate delegate proxies