--- a/src/NSubstitute/Core/MethodFormatter.cs
+++ b/src/NSubstitute/Core/MethodFormatter.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
-using NSubstitute.Proxies.DelegateProxy;
 
 namespace NSubstitute.Core
 {
@@ -15,10 +14,7 @@ namespace NSubstitute.Core
         public string Format(MethodInfo methodInfo, IEnumerable<string> arguments)
         {
             var args = string.Join(", ", arguments);
-
-            return IsDelegateProxy(methodInfo)
-                ? string.Format("Invoke({0})", args)
-                : string.Format("{0}{1}({2})", methodInfo.Name, FormatGenericType(methodInfo), args);
+            return $"{methodInfo.Name}{FormatGenericType(methodInfo)}({args})";
         }
 
         private string FormatGenericType(MethodInfo methodInfoOfCall)
@@ -26,11 +22,6 @@ namespace NSubstitute.Core
             if (!methodInfoOfCall.IsGenericMethod) return string.Empty;
             var genericArgs = methodInfoOfCall.GetGenericArguments();
             return "<" + string.Join(", ", genericArgs.Select(x => x.GetNonMangledTypeName()).ToArray()) + ">";
-        }
-
-        private static bool IsDelegateProxy(MethodInfo methodInfo)
-        {
-            return methodInfo.GetCustomAttribute<ProxiedDelegateTypeAttribute>() != null;
         }
     }
 }