--- a/src/NSubstitute/Proxies/DelegateProxy/ProxiedDelegateTypeAttribute.cs
+++ b/src/NSubstitute/Proxies/DelegateProxy/ProxiedDelegateTypeAttribute.cs
@@ -2,6 +2,7 @@ using System;
 
 namespace NSubstitute.Proxies.DelegateProxy
 {
+    [Obsolete("This class is deprecated and will be removed in future versions of the product.")]
     [AttributeUsage(AttributeTargets.Method)]
     public class ProxiedDelegateTypeAttribute : Attribute
     {