--- a/src/NSubstitute/Proxies/ProxyFactory.cs
+++ b/src/NSubstitute/Proxies/ProxyFactory.cs
@@ -4,6 +4,7 @@ using NSubstitute.Core;
 
 namespace NSubstitute.Proxies
 {
+    [Obsolete("This class is deprecated and will be removed in future versions of the product.")]
     public class ProxyFactory : IProxyFactory
     {
         private readonly IProxyFactory _delegateFactory;
@@ -17,7 +18,7 @@ namespace NSubstitute.Proxies
 
         public object GenerateProxy(ICallRouter callRouter, Type typeToProxy, Type[] additionalInterfaces, object[] constructorArguments)
         {
-            var isDelegate = typeToProxy.GetTypeInfo().IsSubclassOf(typeof(Delegate));
+            var isDelegate = typeToProxy.IsDelegate();
             return isDelegate 
                 ? _delegateFactory.GenerateProxy(callRouter, typeToProxy, additionalInterfaces, constructorArguments)
                 : _dynamicProxyFactory.GenerateProxy(callRouter, typeToProxy, additionalInterfaces, constructorArguments);