--- a/src/NSubstitute/Routing/AutoValues/AutoSubstituteProvider.cs
+++ b/src/NSubstitute/Routing/AutoValues/AutoSubstituteProvider.cs
@@ -17,7 +17,7 @@ namespace NSubstitute.Routing.AutoValues
         public bool CanProvideValueFor(Type type)
         {
             return type.GetTypeInfo().IsInterface
-                || type.GetTypeInfo().IsSubclassOf(typeof(Delegate))
+                || type.IsDelegate()
                 || IsPureVirtualClassWithParameterlessConstructor(type);
         }