… src/NSubstitute/IsReadOnlyAttribute.cs