@zvirja 2019-04-13
@dtchepak Added benchmark results - everything is fine. Please take a look at the latest changes and fi you are fine - feel free to merge the code.

Thanks!